上报功能  前端websocket的配置  echarts的K线图的开发  一个特效的组件  还有blog组件的开发

+++ src/pages/main/MainPages.tsx
@@ -40,9 +40,9 @@ function getItem(
 }
 
 const items: MenuItem[] = [
-    getItem('Option 1', '1', <PieChartOutlined />),
-    getItem('Option 2', '2', <DesktopOutlined />),
-    getItem('Option 3', '3', <ContainerOutlined />),
+    getItem('我的首页', '1', <PieChartOutlined />),
+    getItem('返回主页', '2', <DesktopOutlined />),
+    getItem('图表页', '3', <ContainerOutlined />),
 
     getItem('Navigation One', 'sub1', <MailOutlined />, [
         getItem('Option 5', '5'),
@@ -57,6 +57,21 @@ const items: MenuItem[] = [
         getItem('Submenu', 'sub3', null, [getItem('Option 11', '11'), getItem('Option 12', '12')]),
     ]),
 ];
+// 点击跳转的映射
+const hashList: { [key: string]: string } = {
+"1":'/homePage',
+"2":'/homePage',
+"3":'/dashboard',
+"4":'/',
+"5":'/',
+"6":'/',
+"7":'/',
+"8":'/',
+"9":'/',
+"10":'/',
+"11":'/',
+"12":'/',
+};
 
 // 获取 侧边栏数据
 const useMenu = () => {
@@ -78,17 +93,23 @@ const useMenu = () => {
 const App: React.FC = () => {
     const navigate = useNavigate();
     const dispatch = useDispatch();
+    // 退出登录
     const handleOnclick = async () => {
         try {
             const res = await logoutUser();
             console.log('退出登录的', res)
             // 清除 local storage 或 session storage
-            localStorage.clear();
-            sessionStorage.clear();
-            // redux 重置
-            dispatch(logout());
-            navigate('/login');
+            if (res.successed) {
+                localStorage.clear();
+                sessionStorage.clear();
+                // redux 重置
+                dispatch(logout());
+                navigate('/login');
+            } else {
+
+            }
         } catch (error) {
+            console.log(error, '退出登录错误')
             throw error;
         }
     }
@@ -98,6 +119,18 @@ const App: React.FC = () => {
     } = theme.useToken();
     const itemsMenu = useMenu()
     console.log(itemsMenu, '侧边栏的数据结构')
+    // 
+    const menuClick: MenuProps['onClick'] = (e) => {
+        console.log(e, '菜单的点击事件')
+       
+    }
+    const menuonSelect = (e:any) => {
+        // console.log(e, 'menuonSelect')
+        let targetPage='/';
+        targetPage = hashList[e.key as string]
+        navigate(targetPage);
+        console.log(targetPage,'sssssssssss')
+    }
     return (
         <Layout style={{ minHeight: '100vh' }}>
             {/* 这是左边的菜单栏 */}
@@ -112,6 +145,8 @@ const App: React.FC = () => {
                     theme="dark"
                     inlineCollapsed={collapsed}
                     items={items}
+                    onClick={menuClick}
+                    onSelect={menuonSelect}
                 />
             </Sider>
             {/* 这是右边的头部的两个按钮 */}
